[feat] Implement `optional` validator
This allows consumers to mark fields as optional and skip validation
when the field is not present.

The implementation trades off on performance for consumer ergonomics:
we allow the user to put the "optional" validator anywhere in the list
instead of making them always put it first, but this requires us to
traverse the list of validators twice: once up front to scan for the
presence of the "optional" validator so we can know ahead of time
whether we can skip validation when the field doesn't exist, and again
to iterate over the validators.